refactor: move processCowBreeding to its own file
